% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_functions.R
\name{geocode_address}
\alias{geocode_address}
\title{Function to interface with the TN geocoder. Needs a lot of love.}
\usage{
geocode_address(
  df,
  match_on = c(address = "pt_address", city = "pt_city", region = "pt_state", postal =
    "pt_zip"),
  return_fields = c("Score", "Match_addr", "Subregion", "X", "Y")
)
}
\arguments{
\item{df}{dataframe to use and return with additional columns}

\item{match_on}{Named character vector of columns in the dataframe that map to geocoded elements.
geocoder_name=df_name}

\item{return_fields}{Character vector of fields to return. Check TN geocoder site for options.}
}
\value{
Dataframe with added columns from geocoder
}
\description{
Function to interface with the TN geocoder. Needs a lot of love.
}
