% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{niosh_api_single}
\alias{niosh_api_single}
\title{Send a single line to the NIOSH API}
\usage{
niosh_api_single(i, o, id)
}
\arguments{
\item{i}{The industry string}

\item{o}{The occupation string}

\item{id}{The ID string/numeric}
}
\description{
Send a single line to the NIOSH API
}
