% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_functions.R, R/geocoding_functions.R,
%   R/ggplot.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{email_draft}
\alias{email_send}
\alias{tn_county_to_region}
\alias{tn_geocode_vector}
\alias{tn_geocode_addresses}
\alias{tn_api_outputs}
\alias{tn_api_inputs}
\alias{tn_geocoder_specialcases}
\alias{tn_geocoder_url}
\alias{theme_tn}
\alias{tn_validate_color}
\alias{tn_logo_names}
\alias{tn_palette_names}
\alias{tn_logo_fetch}
\alias{add_tn_logo}
\alias{tn_palette_show}
\alias{tn_color_show}
\alias{tn_pal}
\alias{scale_color_tn}
\alias{scale_fill_tn}
\alias{tn_ggplot_colors_off}
\alias{tn_ggplot_colors_on}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{TNbrandR}{\code{\link[TNbrandR]{add_tn_logo}}, \code{\link[TNbrandR]{scale_color_tn}}, \code{\link[TNbrandR]{scale_fill_tn}}, \code{\link[TNbrandR]{theme_tn}}, \code{\link[TNbrandR]{tn_color_show}}, \code{\link[TNbrandR]{tn_ggplot_colors_off}}, \code{\link[TNbrandR]{tn_ggplot_colors_on}}, \code{\link[TNbrandR]{tn_logo_fetch}}, \code{\link[TNbrandR]{tn_logo_names}}, \code{\link[TNbrandR]{tn_pal}}, \code{\link[TNbrandR]{tn_palette_names}}, \code{\link[TNbrandR]{tn_palette_show}}, \code{\link[TNbrandR]{tn_validate_color}}}

  \item{TNemailR}{\code{\link[TNemailR]{email_draft}}, \code{\link[TNemailR]{email_send}}}

  \item{TNgeocodeR}{\code{\link[TNgeocodeR]{tn_api_inputs}}, \code{\link[TNgeocodeR]{tn_api_outputs}}, \code{\link[TNgeocodeR]{tn_county_to_region}}, \code{\link[TNgeocodeR]{tn_geocode_addresses}}, \code{\link[TNgeocodeR]{tn_geocode_vector}}, \code{\link[TNgeocodeR]{tn_geocoder_specialcases}}, \code{\link[TNgeocodeR]{tn_geocoder_url}}}
}}

