% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\docType{data}
\name{tn_census_shapefiles}
\alias{tn_census_shapefiles}
\title{Shapefiles for each census tract in TN}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1701 rows and 13 columns.
}
\usage{
tn_census_shapefiles
}
\description{
Shapefiles and associated data for census tracts in TN
}
\keyword{datasets}
