% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{tn_clean_date}
\alias{tn_clean_date}
\title{Cleans dates in a variety of formats}
\usage{
tn_clean_date(x, format = "\%m/\%d/\%Y")
}
\arguments{
\item{x}{vector of date-like values}

\item{format}{string recognized by format() to format the date}
}
\value{
Date or formatted string
}
\description{
Cleans dates in a variety of formats
}
