% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvps_api.R
\name{tn_mvps_case_class_codes}
\alias{tn_mvps_case_class_codes}
\title{Valid MVPS case statuses}
\usage{
tn_mvps_case_class_codes()
}
\value{
Vector of MVPS case statuses
}
\description{
Valid MVPS case statuses
}
