% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvps_api.R
\name{tn_mvps_credentials_get}
\alias{tn_mvps_credentials_get}
\title{Return locally saved MVPS API credentials}
\usage{
tn_mvps_credentials_get()
}
\value{
List with necessary fields
}
\description{
Return locally saved MVPS API credentials
}
