% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvps_api.R
\name{tn_mvps_credentials_set}
\alias{tn_mvps_credentials_set}
\title{Save MVPS API credentials locally}
\usage{
tn_mvps_credentials_set(client_id, scope, client_secret, subscription_key)
}
\arguments{
\item{client_id}{Client ID}

\item{scope}{Scope}

\item{client_secret}{Client Secret}

\item{subscription_key}{Subscription Key}
}
\value{
Nothing
}
\description{
Save MVPS API credentials locally
}
