% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvps_api.R
\name{tn_mvps_jwt}
\alias{tn_mvps_jwt}
\title{Generate an MVPS access token for use in other API calls}
\usage{
tn_mvps_jwt()
}
\value{
Access token
}
\description{
You must have previously set your MVPS credentials using tn_mvps_credentials_set()
}
