% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvps_api.R
\name{tn_mvps_linelist}
\alias{tn_mvps_linelist}
\title{Pull in linelist data from the MVPS API}
\usage{
tn_mvps_linelist(
  year = 2024,
  from = 1,
  to = 53,
  eventCode = NULL,
  category = NULL,
  classificationStatus = NULL,
  messageType = NULL,
  reconciliationStatus = NULL
)
}
\arguments{
\item{year}{Integer for year of data to pull}

\item{from}{Integer for starting MMWR Week}

\item{to}{Integer for ending MMWR Week}

\item{eventCode}{Condition codes to pull}

\item{category}{Category of event codes to pull}

\item{classificationStatus}{Limit extract to these case statuses (from tn_mvps_case_class_codes())}

\item{messageType}{Limit extract to these message types (from tn_mvps_message_types())}

\item{reconciliationStatus}{Limit extract by reconciliation status}
}
\value{
Dataframe
}
\description{
You must set the MVPS credentials on your machine using tn_mvps_credentials_set() prior to calling this function.
For "complete" API documentation, see https://developer.cdc.gov/docs/mvps-external-api/documentation
}
