% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_functions.R
\name{tn_niosh_api}
\alias{tn_niosh_api}
\title{Standardize a dataframe of industries and occupations}
\usage{
tn_niosh_api(
  df,
  id_col = NA,
  industry_col = "Industry",
  occupation_col = "Occupation"
)
}
\arguments{
\item{df}{A dataframe with industry and occupation columns}

\item{id_col}{String with the column name for the ID. Optional, will be assigned to row number if absent.}

\item{industry_col}{String with the column name for the industry.}

\item{occupation_col}{String with the column name for the occupation.}
}
\description{
This function outputs a dataframe with an ID column, and codes, tiles, and match scores for industry and occupation.
}
