% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TN_tracts.R
\name{TN_tracts}
\alias{TN_tracts}
\title{Pull ACS data from the census api}
\usage{
TN_tracts(
  state,
  county,
  msa,
  neighborhood1 = NULL,
  neighborhood2 = NULL,
  neighborhood3 = NULL,
  census_key = "7e25a5365676472f0dd97248a68e5288a0e65338"
)
}
\arguments{
\item{state}{a character string containing a 2-digit state fips code}

\item{county}{a character string containing a 3-digit county fips code}

\item{msa}{a character string containing a 5-digit CBSA code. See \href{https://public.opendatasoft.com/explore/dataset/core-based-statistical-areas-cbsas-and-combined-statistical-areas-csas/table/}{here}
for CBSA codes}

\item{neighborhood1}{a character vector containing 6-digit TRACTCE character strings}

\item{neighborhood2}{a character vector containing 6-digit TRACTCE character strings}

\item{neighborhood3}{a character vector containing 6-digit TRACTCE character strings}
}
\value{
A data frame containing neighborhood- and MSA-level data for Gen2 investment memo
}
\description{
This function pulls relevant variables at the census tract and MSA level from
the census API.
}
