% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_functions.R, R/geocoding_functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{email_draft}
\alias{email_send}
\alias{tn_county_to_region}
\alias{tn_geocode_vector}
\alias{tn_geocode_addresses}
\alias{tn_api_outputs}
\alias{tn_api_inputs}
\alias{tn_geocoder_specialcases}
\alias{tn_geocoder_url}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{TNemailR}{\code{\link[TNemailR]{email_draft}}, \code{\link[TNemailR]{email_send}}}

  \item{TNgeocodeR}{\code{\link[TNgeocodeR]{tn_api_inputs}}, \code{\link[TNgeocodeR]{tn_api_outputs}}, \code{\link[TNgeocodeR]{tn_county_to_region}}, \code{\link[TNgeocodeR]{tn_geocode_addresses}}, \code{\link[TNgeocodeR]{tn_geocode_vector}}, \code{\link[TNgeocodeR]{tn_geocoder_specialcases}}, \code{\link[TNgeocodeR]{tn_geocoder_url}}}
}}

